refactor utility function into a utility package
