+++ container/go/cmd/puller/BUILD
@@ -32,6 +32,7 @@ go_library(
     importpath = "github.com/bazelbuild/rules_docker",
     visibility = ["//visibility:private"],
     deps = [
+        "//container/go/pkg/compat:go_default_library",
         "//container/go/pkg/oci:go_default_library",
         "@com_github_google_go_containerregistry//pkg/authn:go_default_library",
         "@com_github_google_go_containerregistry//pkg/name:go_default_library",
